Fix device labeling

+++ autotest/c/src/tests/src/test_gpio.c
@@ -15,7 +15,7 @@ void test_gpio_on(void)
     Status res;
     devh_t dev;
     TEST_START();
-    res = __sys_get_device_handle((uint8_t)devices[DEV_ID_LED0].id);
+    res = __sys_get_device_handle(devices[DEV_ID_LED0].id);
     copy_from_kernel((uint8_t*)&dev, sizeof(devh_t));
     ASSERT_EQ(res, STATUS_OK);
     LOG("handle is %lx", dev);
@@ -32,7 +32,7 @@ void test_gpio_off(void)
     devh_t dev;
 
     TEST_START();
-    res = __sys_get_device_handle((uint8_t)devices[DEV_ID_LED0].id);
+    res = __sys_get_device_handle(devices[DEV_ID_LED0].id);
     copy_from_kernel((uint8_t*)&dev, sizeof(devh_t));
     ASSERT_EQ(res, STATUS_OK);
     LOG("handle is %lx", dev);
@@ -52,7 +52,7 @@ void test_gpio_toggle(void)
     duration.tag = SLEEP_DURATION_ARBITRARY_MS;
     duration.arbitrary_ms = 250; /* 250 ms*/
     TEST_START();
-    res = __sys_get_device_handle((uint8_t)devices[DEV_ID_LED0].id);
+    res = __sys_get_device_handle(devices[DEV_ID_LED0].id);
     copy_from_kernel((uint8_t*)&dev, sizeof(devh_t));
     res = __sys_gpio_configure(dev, 0);
     ASSERT_EQ(res, STATUS_OK);
@@ -70,7 +70,7 @@ void test_gpio_invalid_io(void)
     devh_t dev;
 
     TEST_START();
-    res = __sys_get_device_handle((uint8_t)devices[DEV_ID_LED0].id);
+    res = __sys_get_device_handle(devices[DEV_ID_LED0].id);
     copy_from_kernel((uint8_t*)&dev, sizeof(devh_t));
     res = __sys_gpio_configure(dev, 4);
     ASSERT_EQ(res, STATUS_INVALID);
